test(skippy): own split capacity regressions
